+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v26.1.6-mc26.1.x] - 2026-05-06
+
+### Added
+
+- Add `SimpleHumanoidArmorLayer` and `SimpleItemInHandLayer`
+
+### Fixed
+
+- Fix `EventResultHolder` not accepting a `null` value
+
 ## [v26.1.5-mc26.1.x] - 2026-05-06
 
 ### Added